+++ WebApi/SwaggerUi.cs
@@ -8,7 +8,7 @@ namespace AndreasBehrend.NINA.Phd2Api.WebApi {
   <meta charset=""UTF-8"">
   <meta name=""viewport"" content=""width=device-width, initial-scale=1.0"">
   <title>PHD2 API</title>
-  <link rel=""stylesheet"" href=""https://unpkg.com/swagger-ui-dist@5/swagger-ui.css"">
+  <link rel=""stylesheet"" href=""/api/v1/swagger-assets/swagger-ui.css"">
   <style>
     html {{ box-sizing: border-box; overflow-y: scroll; }}
     *, *:before, *:after {{ box-sizing: inherit; }}
@@ -19,8 +19,8 @@ namespace AndreasBehrend.NINA.Phd2Api.WebApi {
 </head>
 <body>
   <div id=""swagger-ui""></div>
-  <script src=""https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js""></script>
-  <script src=""https://unpkg.com/swagger-ui-dist@5/swagger-ui-standalone-preset.js""></script>
+  <script src=""/api/v1/swagger-assets/swagger-ui-bundle.js""></script>
+  <script src=""/api/v1/swagger-assets/swagger-ui-standalone-preset.js""></script>
   <script>
     window.onload = function() {{
       SwaggerUIBundle({{